+++ AZTU-Akademik/Models/ResearcherLanguage.cs
@@ -13,7 +13,7 @@ namespace AZTU_Akademik.Models
         public int? ResearcherId { get; set; }
         public short? LanguageId { get; set; }
         public short? LevelId { get; set; }
-        public int? FileId { get; set; }
+        public long? FileId { get; set; }
 
         public virtual File File { get; set; }
         public virtual Language Language { get; set; }
